Implement attestations to enable users to skip checks at compile time.
Specifically, we provide the compiler with type information to specify that
an integer is always positive or no greater than a certain threshold. This
can then be used to skip checks in other sanisizer functions for improved
efficiency, e.g., if we get an int that is known to be <= 200, we can skip
any overflow checks when casting to a uint8_t.

We introduce checks for non-negative integers in the majority of functions,
where any negative value causes an an out-of-range error to be raised. This is
safer as the user is no longer responsible for checking for non-negative
values, but if they do know that the values are non-negative, they can skip our
checks by wrapping the input in an appropriate Attestation.

The Attestations also provide an opportunity for optimizing some of the
arithmetic statements when more than 2 arguments are involved, because we can
more accurately propagate the maximum value of the intermediates to subsequent
calculations. This allows more checks to be skipped at compile time.

Also replaced our custom OverflowError with std::overflow_error.